backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "UrbanExplorer AI agent for location discovery"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.7.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-socketio>=5.11.0",
    "python-engineio>=4.9.0",
    "requests>=2.31.0",
    "websocket-client>=1.6.0"
]
